Gradle edits for testing
