+++ build.gradle
@@ -17,26 +17,7 @@ dependencies {
     testCompile 'junit:junit:4.12'
 }
 
-test{
-    systemProperties 'property': 'value'
-    useTestNG()
-    include 'src/chessTests/..*'
-    testLogging.showStandardStreams = true
-
-    // set heap size for the test JVM(s)
-    minHeapSize = "128m"
-    maxHeapSize = "512m"
-
-    // set JVM arguments for the test JVM(s)
-    jvmArgs '-XX:MaxPermSize=256m'
-
-    // listen to events in the test execution lifecycle
-    beforeTest { descriptor ->
-        logger.lifecycle("Running test: " + descriptor)
-    }
-
-    // listen to standard out and standard error of the test JVM(s)
-    onOutput { descriptor, event ->
-        logger.lifecycle("Test: " + descriptor + " produced standard out/err: " + event.message )
-    }
+task testData(type: Test){
+    include '**/SuiteData.*'
+    reports.junit.Xml.destination = "$buildDir/test-resulsts/DataTestSuite"
 }
